% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity_functions.R
\name{get_b_r}
\alias{get_b_r}
\title{Calculate probability of infection upon receiving an infectious bite (b)}
\usage{
get_b_r(ib, b0, b1, ib0, kb)
}
\arguments{
\item{ib}{Pre-erythrocytic immunity level.}

\item{b0}{Maximum probability of infection due to no pre-erythrocytic immunity (0.59).}

\item{b1}{Maximum reduction in probability of infection due to pre-erythrocytic immunity (0.5).}

\item{ib0}{Scale parameter for pre-erythrocytic immunity (43.9).}

\item{kb}{Shape parameter for pre-erythrocytic immunity (2.16).}
}
\description{
Calculate probability of infection upon receiving an infectious bite (b)
}
