% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_functions.R
\name{get_clinical_hazard_no_het}
\alias{get_clinical_hazard_no_het}
\title{Calculate the risk of clinical malaria in the absence of biting heterogeneity}
\usage{
get_clinical_hazard_no_het(infection_hazard, clinical_probability)
}
\arguments{
\item{infection_hazard}{Infection hazard over time.}

\item{clinical_probability}{Probability of developing clinical symptoms.}
}
\description{
Calculate the risk of clinical malaria in the absence of biting heterogeneity
}
