% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_functions.R
\name{get_epsilon}
\alias{get_epsilon}
\title{Calculate age-specific entomological inoculation rate}
\usage{
get_epsilon(age, eir, rho, a0, gamma_llin = 1)
}
\arguments{
\item{age}{Vector of age/time (in days) over a sufficiently long time, e.g. 1:(365 * 12).}

\item{eir}{EIR per day.}

\item{rho}{Age dependent biting parameter (0.85).}

\item{a0}{Age dependent biting parameter (2920 days).}

\item{gamma_llin}{Adjustment to exposure for individuals sleeping under a bednet.}
}
\description{
Calculate age-specific entomological inoculation rate
}
