% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity_functions.R
\name{get_icm}
\alias{get_icm}
\title{Maternal immunity function}
\usage{
get_icm(age, ica20, pm, dm)
}
\arguments{
\item{age}{Vector of simulation age/time (in days) over a sufficiently long period, e.g. 1:(365 * 10). Cannot be longer than vx.}

\item{ica20}{Average level of immunity in a 20 year old women.}

\item{pm}{Maternal immunity relative to mother's immunity level (0.774).}

\item{dm}{Average duration of maternal immunity (67.7 days).}
}
\value{
Maternal immunity over time.
}
\description{
Maternal immunity is a function of the average immunity levels of a 20-year old woman.
}
