% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_functions.R
\name{get_infection_hazard}
\alias{get_infection_hazard}
\title{Calculate the risk of infection}
\usage{
get_infection_hazard(epsilon, b, seasonality = NULL, vaccine_efficacy = NULL)
}
\arguments{
\item{epsilon}{Vector of age-specific EIR.}

\item{b}{Probability of infection upon mosquito bite.}

\item{vaccine_efficacy}{Vaccine efficacy against infection (proportion). Default = NULL - no vaccine.}

\item{season}{Seasonality profile (vector of seasonality per day). Default = NULL - no seasonality. Can also be generated using \code{get_season}.}
}
\description{
Calculate the risk of infection
}
