% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity_functions.R
\name{get_maternal_start}
\alias{get_maternal_start}
\title{Estimate the average level of immunity in a 20 year old women
to inform maternal immunity levels}
\usage{
get_maternal_start(
  eir,
  gamma_llin,
  season,
  rho,
  a0,
  ub,
  db,
  b0,
  b1,
  ib0,
  kb,
  uc,
  dc,
  cpp = TRUE
)
}
\arguments{
\item{eir}{EIR per day.}

\item{gamma_llin}{Adjustment to exposure for individuals sleeping under a bednet. Set to 1 if not adjusting for this.}

\item{season}{Seasonality profile (vector of seasonality per day). Can be generated using \code{get_season}.}

\item{rho}{Age dependent biting parameter (0.85).}

\item{a0}{Age dependent biting parameter (2920 days).}

\item{ub}{Period in which pre-erythrocytic immunity cannot be boosted (7.2 days).}

\item{db}{Average duration of pre-erythrocytic immunity (3650 days).}

\item{b0}{Maximum probability of infection due to no pre-erythrocytic immunity (0.59).}

\item{b1}{Maximum reduction in probability of infection due to pre-erythrocytic immunity (0.5).}

\item{ib0}{Scale parameter for pre-erythrocytic immunity (43.9).}

\item{kb}{Shape parameter for pre-erythrocytic immunity (2.16).}

\item{uc}{Period in which clinical immunity cannot be boosted (6.06 days).}

\item{dc}{Average duration of clinical immunity (10950 days).}

\item{cpp}{Use cpp functions. Default = TRUE.}
}
\description{
Estimate the average level of immunity in a 20 year old women
to inform maternal immunity levels
}
