% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity_functions.R
\name{get_phi_r}
\alias{get_phi_r}
\title{Calculate probability of an infection leading to a clinical malaria episode (phi)}
\usage{
get_phi_r(ica, icm, phi0, phi1, ic0, kc)
}
\arguments{
\item{ica}{Clinical immunity level.}

\item{icm}{Maternal immunity level.}

\item{phi0}{Maximum probability of clinical disease due to no clinical immunity (0.792).}

\item{phi1}{Maximum reduction in probability of clinical disease due to clinical immunity (0.00074)}

\item{ic0}{Scale parameter for clinical immunity (18).}

\item{kc}{Shape parameter for clinical immunity (2.37).}
}
\description{
Calculate probability of an infection leading to a clinical malaria episode (phi)
}
