% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{get_season}
\alias{get_season}
\title{Generate seasonality curve}
\usage{
get_season(t = 1:365, g0, g1, g2, g3, h1, h2, h3, floor)
}
\arguments{
\item{t}{Time (days). Default = 1:365.}

\item{g0}{Fourier series parameter.}

\item{g1}{Fourier series parameter.}

\item{g2}{Fourier series parameter.}

\item{g3}{Fourier series parameter.}

\item{h1}{Fourier series parameter.}

\item{h2}{Fourier series parameter.}

\item{h3}{Fourier series parameter.}

\item{floor}{Minimum value in seasonality curve.}
}
\description{
Generate seasonality curve
}
