% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_functions.R
\name{get_severe_hazard_no_het}
\alias{get_severe_hazard_no_het}
\title{Calculate the risk of severe malaria in the absence of biting heterogeneity}
\usage{
get_severe_hazard_no_het(infection_hazard, severe_probability)
}
\arguments{
\item{infection_hazard}{Infection hazard over time.}

\item{severe_probability}{Probability of developing severe malaria.}
}
\description{
Calculate the risk of severe malaria in the absence of biting heterogeneity
}
