% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity_functions.R
\name{get_theta}
\alias{get_theta}
\title{Calculate probability of an infection leading to a severe malaria episode (theta)}
\usage{
get_theta(age, iva, ivm, theta0, theta1, iv0, kv, fv0, av, gammav)
}
\arguments{
\item{age}{Vector of simulation age/time (in days).}

\item{iva}{Severe immunity level.}

\item{ivm}{Maternal severe immunity level.}

\item{theta0}{Maximum probability of severe malaria due to no severe immunity (0.075).}

\item{theta1}{Maximum reduction in probability of severe malaria due to severe immunity (0.0001)}

\item{iv0}{Scale parameter for severe immunity (1.096).}

\item{kv}{Shape parameter for severe immunity (2.00).}

\item{fv0}{Age-dependent modifier (0.141).}

\item{av}{Age-dependent modifier (2493.41).}

\item{gammav}{Age-dependent modifier (2.91).}
}
\description{
Calculate probability of an infection leading to a severe malaria episode (theta)
}
