% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_functions.R
\name{get_clinical_hazard}
\alias{get_clinical_hazard}
\title{Simulate average clinical hazard for a cohort}
\usage{
get_clinical_hazard(
  eir,
  age_at_enrollment,
  gamma_llin,
  vx,
  season,
  icm,
  age,
  n,
  zeta,
  weight,
  rho,
  a0,
  ub,
  db,
  b0,
  b1,
  ib0,
  kb,
  uc,
  dc,
  phi0,
  phi1,
  ic0,
  kc,
  cpp = TRUE
)
}
\arguments{
\item{eir}{EIR per day.}

\item{age_at_enrollment}{Age at enrollment into trial (days).}

\item{gamma_llin}{Adjustment to exposure for individuals sleeping under a bednet.}

\item{vx}{Vaccine efficacy over time (proportion).}

\item{season}{Seasonality profile (vector of seasonality per day).}

\item{icm}{Maternal immunity level.}

\item{age}{Vector of age/time (in days) over a sufficiently long time, e.g. 1:(365 * 12).}

\item{n}{Number of biting heterogeneity groups.}

\item{zeta}{Heterogeneity parameters.}

\item{weight}{Heterogeneity weights.}

\item{rho}{Age dependent biting parameter (0.85).}

\item{a0}{Age dependent biting parameter (2920 days).}

\item{ub}{Period in which pre-erythrocytic immunity cannot be boosted (7.2 days).}

\item{db}{Average duration of pre-erythrocytic immunity (3650 days).}

\item{b0}{Maximum probability of infection due to no pre-erythrocytic immunity (0.59).}

\item{b1}{Maximum reduction in probability of infection due to pre-erythrocytic immunity (0.5).}

\item{ib0}{Scale parameter for pre-erythrocytic immunity (43.9).}

\item{kb}{Shape parameter for pre-erythrocytic immunity (2.16).}

\item{uc}{Period in which clinical immunity cannot be boosted (6.06 days).}

\item{dc}{Average duration of clinical immunity (10950 days).}

\item{phi0}{Maximum probability of clinical disease due to no clinical immunity (0.792).}

\item{phi1}{Maximum reduction in probability of clinical disease due to clinical immunity (0.00074)}

\item{ic0}{Scale parameter for clinical immunity (18).}

\item{kc}{Shape parameter for clinical immunity (2.37).}

\item{cpp}{Use cpp functions. Default = TRUE}
}
\description{
Simulates clinical incidence in a cohort of children over time since birth (starting at age 0).
Clinical incidence is returned from enrollment into the trial, defined by @param age_at_enrollment.
}
