% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity_functions.R
\name{get_icm}
\alias{get_icm}
\title{Maternal immunity function}
\usage{
get_icm(age, ica20, pm, dm)
}
\arguments{
\item{age}{Vector of age/time (in days) over a sufficiently long time, e.g. 1:(365 * 12).}

\item{ica20}{Average level of immunity in a 20 year old women.}

\item{pm}{Maternal immunity relative to mother's immunity level.}

\item{dm}{Average duration of maternal immunity.}
}
\value{
Maternal immunity over time.
}
\description{
Maternal immunity is a function of the average immunity levels of a 20-year old woman.
}
