% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_functions.R
\name{get_infection_hazard}
\alias{get_infection_hazard}
\title{Calculate the risk of infection}
\usage{
get_infection_hazard(epsilon, b, seasonality = NULL, vaccine_efficacy = NULL)
}
\arguments{
\item{epsilon}{Vector of age-dependent EIR over time/age.}

\item{b}{Probability of infection upon mosquito bite.}

\item{seasonality}{Seasonality profile.}

\item{vaccine_efficacy}{Vaccine efficacy against infection (proportion)}
}
\description{
Calculate the risk of infection
}
